+++ src/ctests/multiplex2.c
@@ -67,7 +67,7 @@ case1( void )
 	do {
 		if ( PAPI_get_event_info( i, &pset ) == PAPI_OK ) {
 			if ( pset.count && ( strcmp( pset.derived, "NOT_DERIVED" ) == 0 ) ) {
-				retval = PAPI_add_event( EventSet, pset.event_code );
+				retval = PAPI_add_event( EventSet, ( int ) pset.event_code );
 				if ( retval != PAPI_OK )
 					test_fail( __FILE__, __LINE__, "PAPI_add_event", retval );
 				else {
